Inline the section debug initializers next to their previews
