+++ Sources/ScoutUI/Home/Section/Alert/HomeAlertSection.swift
@@ -75,6 +75,17 @@ struct HomeAlertSection: View {
     }
 }
 
+#if DEBUG
+    extension HomeAlertSection {
+        init(statuses: [AlertStatus]) {
+            let alerts = AlertProvider()
+            alerts.result = .success(statuses)
+
+            self.init(alerts: alerts, path: .constant([]))
+        }
+    }
+#endif
+
 #Preview {
     NavigationStack {
         InsetList {
