+++ Sources/ScoutUI/Home/Section/Log/HomeLogSection.swift
@@ -51,6 +51,24 @@ struct HomeLogSection: View {
     }
 }
 
+#if DEBUG
+    extension HomeLogSection {
+        init(period: Period) {
+            let log = HomeLogProvider()
+            for value in Period.allCases {
+                log.period = value
+                log.result = .success(HomeLogProvider.sample(for: value))
+            }
+            log.period = period
+
+            let devices = DevicesProvider()
+            devices.result = .success(.sample)
+
+            self.init(period: period, log: log, devices: devices, path: .constant([]))
+        }
+    }
+#endif
+
 #Preview {
     NavigationStack {
         InsetList {
